Migrated the codebase to GLFW and GLAD + Updates to the README file
I modified the main.cpp file, changing the GLEW + FreeGLUT libraries to GLFW + GLAD.
This change was done mainly because of personal preference in the latter combination.

Because of this change I had to implement some new functions (drawSolidSphere and buildCustomShapes) as well as refactor the main function structure.

I updated the readme file, changing the dependency part and making the todo list more specific.